Use TeamPathFinder for UWAI distance computations
... reducing the computational effort of those computations to something like 25% of v0.98. They're still one of the biggest time sinks, and presumably still _the_ biggest in games with far more than 18 civs. That said, a test with 36 civs suggests that scalability is no longer a major issue. [advc.104b]

+++ CvGameCoreDLL/UWAIAgent.h
@@ -194,7 +194,7 @@ public:
 	// Can this civ hurry production somehow? (Slavery, Univ. Suffrage)
 	bool canHurry() const;
 	double buildUnitProb() const;
-	double shipSpeed() const;
+	int shipSpeed() const;
 	/*  period: Build-up over how many turns? Will be adjusted to game speed
 		by this function! */
 	double estimateBuildUpRate(PlayerTypes civId, int period = 10) const;
